+++ internal/bind/params.go
@@ -6,15 +6,16 @@ import (
 	"errors"
 	"fmt"
 	"net/url"
+	"reflect"
 	"sort"
 	"time"
 
 	"github.com/google/uuid"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/params"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )
 
 var (
@@ -23,128 +24,306 @@ var (
 	errMultipleQueryParameters = errors.New("only one query arg *table.QueryParameters allowed")
 )
 
-//nolint:gocyclo,funlen
-func toValue(v interface{}) (_ types.Value, err error) {
-	if valuer, ok := v.(driver.Valuer); ok {
-		v, err = valuer.Value()
-		if err != nil {
-			return nil, fmt.Errorf("ydb: driver.Valuer error: %w", err)
+var (
+	uuidType    = reflect.TypeOf(uuid.UUID{})
+	uuidPtrType = reflect.TypeOf((*uuid.UUID)(nil))
+)
+
+func asUUID(v interface{}) (value.Value, bool) {
+	switch reflect.TypeOf(v) {
+	case uuidType:
+		return value.Uuid(v.(uuid.UUID)), true //nolint:forcetypeassert
+	case uuidPtrType:
+		vv := v.(*uuid.UUID) //nolint:forcetypeassert
+		if vv == nil {
+			return value.NullValue(types.UUID), true
 		}
+
+		return value.OptionalValue(value.Uuid(*(v.(*uuid.UUID)))), true //nolint:forcetypeassert
+	}
+
+	return nil, false
+}
+
+func toType(v interface{}) (_ types.Type, err error) { //nolint:funlen
+	switch x := v.(type) {
+	case bool:
+		return types.Bool, nil
+	case int:
+		return types.Int32, nil
+	case uint:
+		return types.Uint32, nil
+	case int8:
+		return types.Int8, nil
+	case uint8:
+		return types.Uint8, nil
+	case int16:
+		return types.Int16, nil
+	case uint16:
+		return types.Uint16, nil
+	case int32:
+		return types.Int32, nil
+	case uint32:
+		return types.Uint32, nil
+	case int64:
+		return types.Int64, nil
+	case uint64:
+		return types.Uint64, nil
+	case float32:
+		return types.Float, nil
+	case float64:
+		return types.Double, nil
+	case []byte:
+		return types.Bytes, nil
+	case string:
+		return types.Text, nil
+	case [16]byte:
+		return nil, xerrors.Wrap(value.ErrIssue1501BadUUID)
+	case time.Time:
+		return types.Timestamp, nil
+	case time.Duration:
+		return types.Interval, nil
+	default:
+		kind := reflect.TypeOf(x).Kind()
+		switch kind {
+		case reflect.Slice, reflect.Array:
+			v := reflect.ValueOf(x)
+			t, err := toType(reflect.New(v.Type().Elem()).Elem().Interface())
+			if err != nil {
+				return nil, xerrors.WithStackTrace(
+					fmt.Errorf("cannot parse slice item type %T: %w",
+						x, errUnsupportedType,
+					),
+				)
+			}
+
+			return types.NewList(t), nil
+		case reflect.Map:
+			v := reflect.ValueOf(x)
+
+			keyType, err := toType(reflect.New(v.Type().Key()).Interface())
+			if err != nil {
+				return nil, fmt.Errorf("cannot parse %T map key: %w",
+					reflect.New(v.Type().Key()).Interface(), err,
+				)
+			}
+			valueType, err := toType(reflect.New(v.Type().Elem()).Interface())
+			if err != nil {
+				return nil, fmt.Errorf("cannot parse %T map value: %w",
+					v.MapKeys()[0].Interface(), err,
+				)
+			}
+
+			return types.NewDict(keyType, valueType), nil
+		case reflect.Struct:
+			v := reflect.ValueOf(x)
+
+			fields := make([]types.StructField, v.NumField())
+
+			for i := range fields {
+				kk, has := v.Type().Field(i).Tag.Lookup("sql")
+				if !has {
+					return nil, xerrors.WithStackTrace(
+						fmt.Errorf("cannot parse %v as key field of struct: %w",
+							v.Field(i).Interface(), errUnsupportedType,
+						),
+					)
+				}
+				tt, err := toType(v.Field(i).Interface())
+				if err != nil {
+					return nil, xerrors.WithStackTrace(
+						fmt.Errorf("cannot parse %v as values of dict: %w",
+							v.Field(i).Interface(), errUnsupportedType,
+						),
+					)
+				}
+
+				fields[i] = types.StructField{
+					Name: kk,
+					T:    tt,
+				}
+			}
+
+			return types.NewStruct(fields...), nil
+		default:
+			return nil, xerrors.WithStackTrace(
+				fmt.Errorf("%T: %w. Create issue for support new type %s",
+					x, errUnsupportedType, supportNewTypeLink(x),
+				),
+			)
+		}
+	}
+}
+
+//nolint:gocyclo,funlen
+func toValue(v interface{}) (_ value.Value, err error) {
+	if x, ok := asUUID(v); ok {
+		return x, nil
 	}
 
 	switch x := v.(type) {
 	case nil:
-		return types.VoidValue(), nil
+		return value.VoidValue(), nil
+	case value.Value:
+		return x, nil
+	}
+
+	if vv := reflect.ValueOf(v); vv.Kind() == reflect.Pointer {
+		if vv.IsNil() {
+			tt, err := toType(reflect.New(vv.Type().Elem()).Elem().Interface())
+			if err != nil {
+				return nil, xerrors.WithStackTrace(
+					fmt.Errorf("cannot parse type of %T: %w",
+						v, err,
+					),
+				)
+			}
+
+			return value.NullValue(tt), nil
+		}
+
+		vv, err := toValue(vv.Elem().Interface())
+		if err != nil {
+			return nil, xerrors.WithStackTrace(
+				fmt.Errorf("cannot parse value of %T: %w",
+					v, err,
+				),
+			)
+		}
+
+		return value.OptionalValue(vv), nil
+	}
+
+	switch x := v.(type) {
+	case nil:
+		return value.VoidValue(), nil
 	case value.Value:
 		return x, nil
 	case bool:
-		return types.BoolValue(x), nil
-	case *bool:
-		return types.NullableBoolValue(x), nil
+		return value.BoolValue(x), nil
 	case int:
-		return types.Int32Value(int32(x)), nil
-	case *int:
-		if x == nil {
-			return types.NullValue(types.TypeInt32), nil
-		}
-		xx := int32(*x)
-
-		return types.NullableInt32Value(&xx), nil
+		return value.Int32Value(int32(x)), nil
 	case uint:
-		return types.Uint32Value(uint32(x)), nil
-	case *uint:
-		if x == nil {
-			return types.NullValue(types.TypeUint32), nil
-		}
-		xx := uint32(*x)
-
-		return types.NullableUint32Value(&xx), nil
+		return value.Uint32Value(uint32(x)), nil
 	case int8:
-		return types.Int8Value(x), nil
-	case *int8:
-		return types.NullableInt8Value(x), nil
+		return value.Int8Value(x), nil
 	case uint8:
-		return types.Uint8Value(x), nil
-	case *uint8:
-		return types.NullableUint8Value(x), nil
+		return value.Uint8Value(x), nil
 	case int16:
-		return types.Int16Value(x), nil
-	case *int16:
-		return types.NullableInt16Value(x), nil
+		return value.Int16Value(x), nil
 	case uint16:
-		return types.Uint16Value(x), nil
-	case *uint16:
-		return types.NullableUint16Value(x), nil
+		return value.Uint16Value(x), nil
 	case int32:
-		return types.Int32Value(x), nil
-	case *int32:
-		return types.NullableInt32Value(x), nil
+		return value.Int32Value(x), nil
 	case uint32:
-		return types.Uint32Value(x), nil
-	case *uint32:
-		return types.NullableUint32Value(x), nil
+		return value.Uint32Value(x), nil
 	case int64:
-		return types.Int64Value(x), nil
-	case *int64:
-		return types.NullableInt64Value(x), nil
+		return value.Int64Value(x), nil
 	case uint64:
-		return types.Uint64Value(x), nil
-	case *uint64:
-		return types.NullableUint64Value(x), nil
+		return value.Uint64Value(x), nil
 	case float32:
-		return types.FloatValue(x), nil
-	case *float32:
-		return types.NullableFloatValue(x), nil
+		return value.FloatValue(x), nil
 	case float64:
-		return types.DoubleValue(x), nil
-	case *float64:
-		return types.NullableDoubleValue(x), nil
+		return value.DoubleValue(x), nil
 	case []byte:
-		return types.BytesValue(x), nil
-	case *[]byte:
-		return types.NullableBytesValue(x), nil
+		return value.BytesValue(x), nil
 	case string:
-		return types.TextValue(x), nil
-	case *string:
-		return types.NullableTextValue(x), nil
+		return value.TextValue(x), nil
 	case []string:
-		items := make([]types.Value, len(x))
+		items := make([]value.Value, len(x))
 		for i := range x {
-			items[i] = types.TextValue(x[i])
+			items[i] = value.TextValue(x[i])
 		}
 
-		return types.ListValue(items...), nil
+		return value.ListValue(items...), nil
+	case value.UUIDIssue1501FixedBytesWrapper:
+		return value.UUIDWithIssue1501Value(x.AsBytesArray()), nil
 	case [16]byte:
 		return nil, xerrors.Wrap(value.ErrIssue1501BadUUID)
-	case *[16]byte:
-		return nil, xerrors.Wrap(value.ErrIssue1501BadUUID)
-	case types.UUIDBytesWithIssue1501Type:
-		return types.UUIDWithIssue1501Value(x.AsBytesArray()), nil
-	case *types.UUIDBytesWithIssue1501Type:
-		if x == nil {
-			return types.NullableUUIDValueWithIssue1501(nil), nil
-		}
-		val := x.AsBytesArray()
-
-		return types.NullableUUIDValueWithIssue1501(&val), nil
-	case uuid.UUID:
-		return types.UuidValue(x), nil
-	case *uuid.UUID:
-		return types.NullableUUIDTypedValue(x), nil
 	case time.Time:
-		return types.TimestampValueFromTime(x), nil
-	case *time.Time:
-		return types.NullableTimestampValueFromTime(x), nil
+		return value.TimestampValueFromTime(x), nil
 	case time.Duration:
-		return types.IntervalValueFromDuration(x), nil
-	case *time.Duration:
-		return types.NullableIntervalValueFromDuration(x), nil
+		return value.IntervalValueFromDuration(x), nil
 	default:
-		return nil, xerrors.WithStackTrace(
-			fmt.Errorf("%T: %w. Create issue for support new type %s",
-				x, errUnsupportedType, supportNewTypeLink(x),
-			),
-		)
+		kind := reflect.TypeOf(x).Kind()
+		switch kind {
+		case reflect.Slice, reflect.Array:
+			v := reflect.ValueOf(x)
+			list := make([]value.Value, v.Len())
+
+			for i := range list {
+				list[i], err = toValue(v.Index(i).Interface())
+				if err != nil {
+					return nil, xerrors.WithStackTrace(
+						fmt.Errorf("cannot parse %d item of slice %T: %w",
+							i, x, err,
+						),
+					)
+				}
+			}
+
+			return value.ListValue(list...), nil
+		case reflect.Map:
+			v := reflect.ValueOf(x)
+			fields := make([]value.DictValueField, 0, len(v.MapKeys()))
+			iter := v.MapRange()
+			for iter.Next() {
+				kk, err := toValue(iter.Key().Interface())
+				if err != nil {
+					return nil, fmt.Errorf("cannot parse %v map key: %w",
+						iter.Key().Interface(), err,
+					)
+				}
+				vv, err := toValue(iter.Value().Interface())
+				if err != nil {
+					return nil, fmt.Errorf("cannot parse %v map value: %w",
+						iter.Value().Interface(), err,
+					)
+				}
+				fields = append(fields, value.DictValueField{
+					K: kk,
+					V: vv,
+				})
+			}
+
+			return value.DictValue(fields...), nil
+		case reflect.Struct:
+			v := reflect.ValueOf(x)
+
+			fields := make([]value.StructValueField, v.NumField())
+
+			for i := range fields {
+				kk, has := v.Type().Field(i).Tag.Lookup("sql")
+				if !has {
+					return nil, xerrors.WithStackTrace(
+						fmt.Errorf("cannot parse %q as key field of struct: %w",
+							v.Type().Field(i).Name, errUnsupportedType,
+						),
+					)
+				}
+				vv, err := toValue(v.Field(i).Interface())
+				if err != nil {
+					return nil, xerrors.WithStackTrace(
+						fmt.Errorf("cannot parse %v as values of dict: %w",
+							v.Index(i).Interface(), err,
+						),
+					)
+				}
+
+				fields[i] = value.StructValueField{
+					Name: kk,
+					V:    vv,
+				}
+			}
+
+			return value.StructValue(fields...), nil
+		default:
+			return nil, xerrors.WithStackTrace(
+				fmt.Errorf("%T: %w. Create issue for support new type %s",
+					x, errUnsupportedType, supportNewTypeLink(x),
+				),
+			)
+		}
 	}
 }
 
@@ -159,13 +338,6 @@ func supportNewTypeLink(x interface{}) string {
 
 func toYdbParam(name string, value interface{}) (*params.Parameter, error) {
 	if na, ok := value.(driver.NamedValue); ok {
-		n, v := na.Name, na.Value
-		if n != "" {
-			name = n
-		}
-		value = v
-	}
-	if na, ok := value.(sql.NamedArg); ok {
 		n, v := na.Name, na.Value
 		if n != "" {
 			name = n