+++ topic/topictypes/topictypes_test.go
@@ -152,6 +152,7 @@ func TestTopicDescriptionFromRaw(t *testing.T) {
 	}
 
 	for _, v := range testData {
+		v := v
 		t.Run(
 			v.testName, func(t *testing.T) {
 				d := TopicDescription{}
@@ -321,6 +322,7 @@ func TestTopicConsumerDescriptionFromRaw(t *testing.T) {
 		},
 	}
 	for _, v := range testData {
+		v := v
 		t.Run(
 			v.testName, func(t *testing.T) {
 				d := TopicConsumerDescription{}